rivalradar/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
groq
httpx
beautifulsoup4
python-dotenv
pydantic
requests
lxml
playwright
pytest
pytest-asyncio
